make the docs nazi happy
