+++ src/lib.rs
@@ -1,7 +1,7 @@
 //! `crdts` is a library of thoroughly-tested, serializable CRDT's
 //! ported from the riak_dt library to rust.
 #![crate_type = "lib"]
-// #![deny(missing_docs)]
+#![deny(missing_docs)]
 
 pub use error::{Result, Error};
 pub use gcounter::GCounter;
