+++ src/map.rs
@@ -281,6 +281,7 @@ impl<K: Key, V: Val<A>, A: Actor> Map<K, V, A> {
          }
     }
 
+    /// Returns the next dot for the given actor
     pub fn dot(&self, actor: impl Into<A>) -> Dot<A> {
         let actor = actor.into();
         let counter = self.clock.get(&actor) + 1;
